==> notebooks/2-mf-preprocessing-draft.ipynb <==
## Objectives :

- Only important features and target in the dataset
- Label encoding for categorical feature (only `track_genre`, and `explicit`)
- Normalization for numerical features

import pandas as pd
from pycparser.c_ast import Label

dataset_path = "../data/raw/dataset.csv"
# Load data
df = pd.read_csv(dataset_path, low_memory=False)

# Columns
df.columns

from sklearn.preprocessing import LabelEncoder

label_encoder = LabelEncoder()

df['track_genre_encoded'] = label_encoder.fit_transform(df['track_genre'])

df['track_genre_encoded'].value_counts()

df['explicit'] = df['explicit'].replace({True: 1, False: 0})

# Drop missing values
df.dropna(inplace=True)

df.shape

# Drop columns : 'Unnamed: 0', 'track_id', 'artists', 'album_name', 'track_name'

columns_to_drop = ['Unnamed: 0', 'track_id', 'artists', 'album_name', 'track_name']

df.drop(columns=columns_to_drop, inplace=True)

df.drop(columns=['track_genre'], inplace=True)

df.head()

# Save the preprocessed dataset
preprocessed_dataset_path = "../data/processed/preprocessed_dataset.csv"
df.to_csv(preprocessed_dataset_path, index=False)
