==> notebooks/3-mf-making-dataset.ipynb <==
import pandas as pd

# Load data
df = pd.read_csv('../data/preprocessed/preprocessed_dataset.csv')

df.head()

col_to_standardize = ['popularity', 'duration_ms', 'explicit', 'danceability', 'energy', 'loudness', 'speechiness',
                      'acousticness', 'instrumentalness', 'liveness', 'valence', 'tempo']

# Standardize with sklearn
from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()

df[col_to_standardize] = scaler.fit_transform(df[col_to_standardize])

# Shape
df.shape

df.describe()

# Save the dataset
df.to_csv('../data/processed/dataset_ready.csv', index=False)
