==> notebooks/4-mf-run-sagemaker-locally.ipynb <==
import sagemaker
import boto3


# Session in eu-west-3
sagemaker_session = sagemaker.Session(boto3.Session(region_name='eu-west-3'))

role = "arn:aws:iam::476114139467:role/service-role/AmazonSageMakerServiceCatalogProductsUseRole"

# S3 bucket in eu-west-3
train_input = 's3://s3-melodai-project/processed/83/ab60d273bba7b233b418a0942de534'

# S3 output in eu-west-3 for model
output_path = 's3://s3-melodai-project/models'

# Sagemaker SKlearn Estimator
from sagemaker.sklearn.estimator import SKLearn

# Define the SKLearn Estimator (Random Forest)
sklearn = SKLearn(
    entry_point='/home/matthieu/UbuntuData/PycharmProjects/MelodAI/src/models/train_model.py',
    role=role,
    instance_count=1,
    instance_type='ml.t3.medium',
    framework_version='0.23-1',
    sagemaker_session=sagemaker_session,
    output_path=output_path
)

# Train the model
sklearn.fit({'train': train_input})
